Added actions and reducers to load games from API
