+++ src/actions/index.ts
@@ -2,3 +2,4 @@ export * from './game';
 export * from './types';
 export * from './settings';
 export * from './navigation';
+export * from './games';
