+++ src/components/App/App.tsx
@@ -11,49 +11,49 @@ import { Menu } from '../Menu/Menu';
 import { About } from '../About/About';
 import { Settings } from '../Settings/Settings';
 
-const games = {
-    easy1: [
-        [null, null, null, null, 9, null, 8, 7, 2],
-        [7, null, 2, null, 8, 1, 4, 5, null],
-        [9, null, null, 2, null, null, null, 1, 3],
-        [1, null, null, 9, null, null, null, 4, null],
-        [null, 7, null, 1, null, 5, null, null, null],
-        [4, 2, null, null, null, 8, 5, 6, null],
-        [null, 3, 5, null, 8, null, null, null, null],
-        [8, null, 4, null, 3, 6, 5, null, null],
-        [null, 9, 6, 7, null, null, null, 3, 2],
-    ],
-    almostFinished1: [
-        [5, 1, 3, 6, 9, 4, 8, 7, 2],
-        [7, 6, 2, 3, 8, 1, 4, 5, 9],
-        [null, null, 4, 2, 5, 7, 6, 1, 3],
-        [1, 5, 8, 9, 2, 6, 3, 4, 7],
-        [6, 7, 3, 1, 4, 5, 2, 9, 8],
-        [4, 2, 9, 3, 7, 8, 5, 6, 1],
-        [7, 3, 5, 2, 8, 1, 4, 6, 9],
-        [8, 2, 4, 9, 3, 6, 5, 1, 7],
-        [1, 9, 6, 7, 4, 5, 8, 3, 2]
-    ],
-};
+// const games = {
+//     easy1: [
+//         [null, null, null, null, 9, null, 8, 7, 2],
+//         [7, null, 2, null, 8, 1, 4, 5, null],
+//         [9, null, null, 2, null, null, null, 1, 3],
+//         [1, null, null, 9, null, null, null, 4, null],
+//         [null, 7, null, 1, null, 5, null, null, null],
+//         [4, 2, null, null, null, 8, 5, 6, null],
+//         [null, 3, 5, null, 8, null, null, null, null],
+//         [8, null, 4, null, 3, 6, 5, null, null],
+//         [null, 9, 6, 7, null, null, null, 3, 2],
+//     ],
+//     almostFinished1: [
+//         [5, 1, 3, 6, 9, 4, 8, 7, 2],
+//         [7, 6, 2, 3, 8, 1, 4, 5, 9],
+//         [null, null, 4, 2, 5, 7, 6, 1, 3],
+//         [1, 5, 8, 9, 2, 6, 3, 4, 7],
+//         [6, 7, 3, 1, 4, 5, 2, 9, 8],
+//         [4, 2, 9, 3, 7, 8, 5, 6, 1],
+//         [7, 3, 5, 2, 8, 1, 4, 6, 9],
+//         [8, 2, 4, 9, 3, 6, 5, 1, 7],
+//         [1, 9, 6, 7, 4, 5, 8, 3, 2]
+//     ],
+// };
 
-const solutions = {
-    solutionEasy1: [
-        [5, 1, 3, 6, 9, 4, 8, 7, 2],
-        [7, 6, 2, 3, 8, 1, 4, 5, 9],
-        [9, 8, 4, 2, 5, 7, 6, 1, 3],
-        [1, 5, 8, 9, 2, 6, 3, 4, 7],
-        [6, 7, 3, 1, 4, 5, 2, 9, 8],
-        [4, 2, 9, 3, 7, 8, 5, 6, 1],
-        [7, 3, 5, 2, 8, 1, 4, 6, 9],
-        [8, 2, 4, 9, 3, 6, 5, 1, 7],
-        [1, 9, 6, 7, 4, 5, 8, 3, 2]
-    ],
-};
+// const solutions = {
+//     solutionEasy1: [
+//         [5, 1, 3, 6, 9, 4, 8, 7, 2],
+//         [7, 6, 2, 3, 8, 1, 4, 5, 9],
+//         [9, 8, 4, 2, 5, 7, 6, 1, 3],
+//         [1, 5, 8, 9, 2, 6, 3, 4, 7],
+//         [6, 7, 3, 1, 4, 5, 2, 9, 8],
+//         [4, 2, 9, 3, 7, 8, 5, 6, 1],
+//         [7, 3, 5, 2, 8, 1, 4, 6, 9],
+//         [8, 2, 4, 9, 3, 6, 5, 1, 7],
+//         [1, 9, 6, 7, 4, 5, 8, 3, 2]
+//     ],
+// };
 
-const game = {
-    start: games.easy1,
-    solution: solutions.solutionEasy1
-} as GameDataType;
+// const game = {
+//     start: games.easy1,
+//     solution: solutions.solutionEasy1
+// } as GameDataType;
 
 interface AppProps extends AppStateToProps {
     setGameStatus: typeof setGameStatus;
@@ -107,6 +107,8 @@ class AppComponent extends React.Component<AppProps, AppState> {
             ? <Menu menuItems={menuItems} />
             : this.renderLinkBack();
 
+        const game = this.props.games[Math.floor(Math.random() * this.props.games.length)];
+
         return (
             <div id="app">
                 <TopBar>{menu}</TopBar>
@@ -130,6 +132,7 @@ interface AppStateToProps {
     gameErrors: number;
     gameTime: Date;
     navigation: Pages;
+    games: GameDataType[];
 }
 
 const mapStateToProps = (store: StoreState,): AppStateToProps => {
@@ -139,7 +142,8 @@ const mapStateToProps = (store: StoreState,): AppStateToProps => {
         gameStatus: status,
         gameErrors: errorCounter,
         gameTime: time,
-        navigation: store.navigation
+        navigation: store.navigation,
+        games: store.games
     };
 };
