+++ src/components/Home/Home.tsx
@@ -1,16 +1,24 @@
 import React from "react";
 import { connect } from "react-redux";
-import { setPage } from "../../actions";
+import { setPage, loadGames } from "../../actions";
 import { StoreState } from "../../reducers";
-import { Pages } from "../../types";
+import { GameDataType, Pages } from "../../types";
 import './Home.scss';
 
 interface HomeProps extends HomeStateToProps {
     setPage: typeof setPage;
+    loadGames: typeof loadGames;
 }
 
 class HomeComponent extends React.Component<HomeProps> {
+    loadGames(): void {
+        if (!this.props.games.length) {
+            this.props.loadGames();
+        }
+    }
+
     onNewGameClick = () => {
+        this.loadGames();
         this.props.setPage(Pages.Game);
     };
 
@@ -28,15 +36,17 @@ class HomeComponent extends React.Component<HomeProps> {
 
 interface HomeStateToProps {
     navigation: Pages;
+    games: GameDataType[];
 }
 
 const mapStateToProps = (store: StoreState): HomeStateToProps => {
     return {
-        navigation: store.navigation
+        navigation: store.navigation,
+        games: store.games
     };
 };
 
 export const Home = connect(
     mapStateToProps,
-    { setPage }
+    { setPage, loadGames }
 )(HomeComponent);
