+++ src/reducers/index.ts
@@ -1,6 +1,7 @@
 import { combineReducers } from "redux";
-import { Pages, SettingsType } from "../types";
+import { GameDataType, Pages, SettingsType } from "../types";
 import { gameReducer, GameState } from "./game";
+import { gamesReducer } from "./games";
 import { navigationReducer } from "./navigation";
 import { settingsReducer } from "./settings";
 
@@ -8,13 +9,18 @@ export interface StoreState {
     game: GameState;
     settings: SettingsType;
     navigation: Pages;
+    games: GameDataType[];
 }
 
 export const reducers = combineReducers<StoreState>({
     game: gameReducer,
     settings: settingsReducer,
-    navigation: navigationReducer
+    navigation: navigationReducer,
+    games: gamesReducer
 });
 
 export * from './game';
 export * from './selectedCell';
+export * from './navigation';
+export * from './settings';
+export * from './games';
