+++ src/types/cell.ts
@@ -1,5 +1,3 @@
-import { SudokuValuesType } from ".";
-
 export type SudokuNumbersType = 1 | 2 | 3 | 4 | 5 | 6 | 7 | 8 | 9;
 
 /**
@@ -48,8 +46,3 @@ export type SudokuSolutionType = [
     CellGroupSolution,
     CellGroupSolution
 ];
-
-export interface GameDataType {
-    start: SudokuValuesType;
-    solution: SudokuSolutionType;
-}
