+++ src/types/games.ts
@@ -1,0 +1,6 @@
+import { SudokuSolutionType, SudokuValuesType } from ".";
+
+export interface GameDataType {
+    start: SudokuValuesType;
+    solution: SudokuSolutionType;
+}
