+++ src/types/index.ts
@@ -4,3 +4,4 @@ export * from './cell';
 export * from './cellGroup';
 export * from './sudoku';
 export * from './navigation';
+export * from './games';
